Resolve "etap7"

Closes #48

See merge request kp.marczynski/dietify!20

+++ src/main/java/pl/marczynski/dietify/recipes/service/impl/RecipeServiceImpl.java
@@ -1,5 +1,6 @@
 package pl.marczynski.dietify.recipes.service.impl;
 
+import javassist.NotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.cache.Cache;
@@ -8,10 +9,14 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import pl.marczynski.dietify.core.domain.User;
+import pl.marczynski.dietify.core.service.UserService;
+import pl.marczynski.dietify.core.web.rest.errors.OperationNotAllowedForCurrentUserException;
 import pl.marczynski.dietify.recipes.domain.Recipe;
 import pl.marczynski.dietify.recipes.repository.RecipeRepository;
 import pl.marczynski.dietify.recipes.service.RecipeService;
 
+import java.time.LocalDate;
 import java.util.Optional;
 
 /**
@@ -25,10 +30,12 @@ public class RecipeServiceImpl implements RecipeService {
 
     private final CacheManager cacheManager;
     private final RecipeRepository recipeRepository;
+    private final UserService userService;
 
-    public RecipeServiceImpl(CacheManager cacheManager, RecipeRepository recipeRepository) {
+    public RecipeServiceImpl(CacheManager cacheManager, RecipeRepository recipeRepository, UserService userService) {
         this.cacheManager = cacheManager;
         this.recipeRepository = recipeRepository;
+        this.userService = userService;
     }
 
     /**
@@ -40,8 +47,25 @@ public class RecipeServiceImpl implements RecipeService {
     @Override
     public Recipe save(Recipe recipe) {
         log.debug("Request to save Recipe : {}", recipe);
+        if (!hasRightsToPersistRecipe(recipe)) {
+            throw new OperationNotAllowedForCurrentUserException();
+        }
         this.clearRecipeCaches(recipe);
-        return recipeRepository.save(recipe);
+        if (recipe.getAuthorId() == null) {
+            recipe.setAuthorId(userService.getCurrentUser().get().getId());
+            recipe.creationDate(LocalDate.now());
+        }
+        recipe.lastEditDate(LocalDate.now());
+        return recipeRepository.saveAndFlush(recipe);
+    }
+
+    private boolean hasRightsToPersistRecipe(Recipe recipe) {
+        Optional<User> currentUser = userService.getCurrentUser();
+        if (!currentUser.isPresent()) {
+            return false;
+        } else if (recipe.getId() == null || recipe.getAuthorId() == null) {
+            return true;
+        } else return recipe.getAuthorId().equals(currentUser.get().getId());
     }
 
     /**
@@ -86,10 +110,22 @@ public class RecipeServiceImpl implements RecipeService {
      * @param id the id of the entity
      */
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws NotFoundException {
         log.debug("Request to delete Recipe : {}", id);
+        Optional<Recipe> recipe = recipeRepository.findOneWithEagerRelationships(id);
+        if (!recipe.isPresent()) {
+            throw new NotFoundException("Recipe not found");
+        }
+        if (!hasRightsToPersistRecipe(recipe.get())) {
+            throw new OperationNotAllowedForCurrentUserException();
+        }
         this.clearRecipeCaches(id);
         recipeRepository.deleteById(id);
+    }
+
+    @Override
+    public Page<Recipe> findByNameContaining(String searchPhrase, Pageable pageable) {
+        return this.recipeRepository.findByNameContainingIgnoreCase(searchPhrase, pageable);
     }
 
     private void clearRecipeCaches(Recipe recipe) {
